Fix issue23

+++ trunk/includes/controllers/SiteController.php
@@ -68,37 +68,44 @@ class SiteController extends BaseController
             $tips=sprintf(ZFramework::t('CONFIG_FILE_NOTEXISTS',array(),$language),CONFIGFILE);
         elseif(!is_writable(CONFIGFILE))
             $tips=sprintf(ZFramework::t('CONFIG_FILE_NOTWRITABLE',array(),$language),CONFIGFILE);
-        //elseif()
+        elseif(!is_writable(APPROOT.'/data/'))
+            $tips= ZFramework::t ('DATADIR_NOT_WRITABLE', array(), $language);
+        elseif(!is_writable(APPROOT.'/plugins/'))
+            $tips= ZFramework::t ('PLUGINDIR_NOT_WRITABLE', array(), $language);
         if(!empty ($_POST['adminname']) && !empty($_POST['adminpass']) && !empty ($_POST['dbname']) && strlen(trim($_POST['adminname']))>2 ){
             $adminname=ZFramework::maple_quotes($_POST['adminname']);
             $adminpass=ZFramework::maple_quotes($_POST['adminpass']);
             $dbname=  ZFramework::maple_quotes($_POST['dbname']);
-            $adminnameString="\n\$admin='$adminname';";
-            $adminpassString="\n\$password='$adminpass';";
-            $dbnameString="\n\$dbname='$dbname';";
-            file_put_contents(CONFIGFILE, $adminnameString,FILE_APPEND);
-            file_put_contents(CONFIGFILE, $adminpassString,FILE_APPEND);
-            file_put_contents(CONFIGFILE, $dbnameString,FILE_APPEND);
-            if(!$this->_model->create_db($dbname)){
-                die ($this->_model->error());
-            }
-            $this->_model->select_db($dbname);
-
-            $tables=array(MESSAGETABLE,  REPLYTABLE,  BADIPTABLE, USERTABLE);
-            $fields=array(
-                        array(array('name'=>'id','auto_increment'=>true),array('name'=>'user'),array('name'=>'content'),array('name'=>'time'),array('name'=>'ip')),
-                        array(array('name'=>'id'),array('name'=>'reply_content'),array('name'=>'reply_time')),
-                        array(array('name'=>'ip')),
-                        array(array('name'=>'uid','auto_increment'=>true),array('name'=>'user'),array('name'=>'pwd'),array('name'=>'email')),
-                        );
-            for($i=0,$t=count($tables);$i<$t;$i++){
-                if(!$this->_model->create_table($tables[$i],$fields[$i])){
-                    die($this->_model->error());
+            if($this->_model->_db_exists($dbname)){
+                $formError=ZFramework::t('APP_DB_EXISTS', array('{db}'=>$dbname), $language);
+            }else{
+                $adminnameString="\n\$admin='$adminname';";
+                $adminpassString="\n\$password='$adminpass';";
+                $dbnameString="\n\$dbname='$dbname';";
+                file_put_contents(CONFIGFILE, $adminnameString,FILE_APPEND);
+                file_put_contents(CONFIGFILE, $adminpassString,FILE_APPEND);
+                file_put_contents(CONFIGFILE, $dbnameString,FILE_APPEND);
+                if(!$this->_model->create_db($dbname)){
+                    die ($this->_model->error());
                 }
+                $this->_model->select_db($dbname);
+
+                $tables=array(MESSAGETABLE,  REPLYTABLE,  BADIPTABLE, USERTABLE);
+                $fields=array(
+                            array(array('name'=>'id','auto_increment'=>true),array('name'=>'user'),array('name'=>'content'),array('name'=>'time'),array('name'=>'ip')),
+                            array(array('name'=>'id'),array('name'=>'reply_content'),array('name'=>'reply_time')),
+                            array(array('name'=>'ip')),
+                            array(array('name'=>'uid','auto_increment'=>true),array('name'=>'user'),array('name'=>'pwd'),array('name'=>'email')),
+                            );
+                for($i=0,$t=count($tables);$i<$t;$i++){
+                    if(!$this->_model->create_table($tables[$i],$fields[$i])){
+                        die($this->_model->error());
+                    }
+                }
+                $newData=array(NULL,$_POST['adminname'],'Welcome to MapleLeaf.:)',time(), getIp());
+                $this->_model->insert(MESSAGETABLE, $newData);
+                $installed=TRUE;
             }
-	    $newData=array(NULL,$_POST['adminname'],'Welcome to MapleLeaf.:)',time(), getIp());
-	    $this->_model->insert(MESSAGETABLE, $newData);
-	    $installed=TRUE;
         }
 	if(file_exists(dirname(dirname(__FILE__)).'/install.php')){
 	    include dirname(dirname(__FILE__)).'/install.php';
